+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [unreleased]
+### Changed
+- Structure of parsed entities, to reflect documents saved in the database (#20)
 ### Added
 - Use the `rust-htslib` for reading VCF files (#1)
 - First Variant model and parsing modules (#2)
@@ -20,5 +22,6 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Infer complete coordinate data (#17)
 - Parse cytobands and collect cytoband start and end for each variant (#18)
 - Parse genmod genetic models (#19)
+- Parsing specific variant fields: `str_swegen_mean` and `str_swegen_std` (STRs) and `somatic_score` (cancer and cancer_sv variants) (#20)
 ### Fixed
 - Normalize chromosome names by stripping the chr prefix (#4)